modifying subgraph

+++ Attacks/Utils/data_utils.py
@@ -228,8 +228,8 @@ def percentage_pos(node, graph):
     mask[dst.unique().long()] = 1
     index = get_index_by_value(a=mask, val=1)
     nodes_id = frontier.nodes()[index]
-    num_pos = graph.ndata['pos_mask'][nodes_id].sum()
-    num_neg = graph.ndata['neg_mask'][nodes_id].sum()
+    num_pos = graph.ndata['pos_mask'][nodes_id.long()].sum()
+    num_neg = graph.ndata['neg_mask'][nodes_id.long()].sum()
     pos_percentage = num_pos.item() / (num_pos.item() + num_neg.item() + 1e-12)
     return pos_percentage
